+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Added missing parentheses in v-for loops
 - StatisticDashboard:
   - Statistics and Regions are sorted alphabetically.
+  - The information text will show under legend when there are more than one statistics selected.
 - LayerPreview for WMTS uses the crs of the service.
 - Issue #1192: Searchbar: The TopicTree search interface supports the search for folders.